동적쿼리 지역 선택 완료, 화면봤으니까 다시 service 구현중

+++ milestone/src/main/java/com/app/milestone/domain/MoneyDTO.java
@@ -21,7 +21,7 @@ import javax.persistence.ManyToOne;
 public class MoneyDTO {
     private String userName;
     private Long userId;
-    private int moneyCash;
+    private Long moneyCash;
 
     public Money toEntity() {
         return Money.builder()
@@ -30,7 +30,7 @@ public class MoneyDTO {
     }
 
     @QueryProjection
-    public MoneyDTO(String userName, Long userId, int moneyCash) {
+    public MoneyDTO(String userName, Long userId, Long moneyCash) {
         this.userName = userName;
         this.userId = userId;
         this.moneyCash = moneyCash;
